pyproject.toml
[project]
name = "chess-gemma"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "chess>=1.11.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
